% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eoption.R
\name{Eoption}
\alias{Eoption}
\title{calculate the price of European option by diffrence method}
\usage{
Eoption(D, sigma, S0, r, q, K, type)
}
\arguments{
\item{D}{expiration date}

\item{sigma}{Annual volatility}

\item{S0}{current price of underlying assets}

\item{r}{Continuously compounded risk-free interest rate}

\item{q}{Bonus rate}

\item{K}{strike price}

\item{type}{call option:"C" or "c"; put option: "p" or "P"}
}
\value{
current price of the American option price
}
\description{
calculate the price of European option by diffrence method
}
\examples{
Eoption(30,0.1,1,0.1,0.1,1,"c")
}
