\name{StatComp21090-package}
\alias{StatComp21090-package}
\docType{package}
\title{
  This package provides functions to calculate option price
}
\description{
  Construct functions to calculate European/Amrican call/put option price by diffrence method(\code{Eoption} and \code{Aoption}). Use this to varify the Option parity formula, and show that the American call option's price is equal to the European.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Bessie

Maintainer: Bessie <zhoubeixi@mail.ustc.edu.cn>
}
\references{
  <Mathematical Models and Methods of Option Pricing (Second Edition)> Lishang Jiang
}
\keyword{ package }
